LayoutTests/fast/animation/request-animation-frame-throttling-aggressiveThermalMitigation-expected.txt

Test that requestAnimationFrame gets throttled when aggressive thermal mitigation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS framesPerSecond < 36 is true
PASS successfullyParsed is true

TEST COMPLETE
